+++ lib/design/atomic/molecules/content_card_description/announcements_content_description.dart
@@ -1,0 +1,151 @@
+import 'package:flutter/material.dart';
+import 'package:stoyco_partners_shared/design/atomic/atoms/content_stat/content_stat.dart';
+import 'package:stoyco_partners_shared/design/atomic/templates/content_card_description.dart';
+import 'package:stoyco_partners_shared/design/models/content_cards/announcements_content_model.dart';
+import 'package:stoyco_partners_shared/design/responsive/gutter.dart';
+import 'package:stoyco_partners_shared/design/responsive/screen_size/stoyco_screen_size.dart';
+import 'package:stoyco_partners_shared/design/utils/foundations/color_foundation.dart';
+import 'package:stoyco_partners_shared/design/utils/tokens/gen/assets.gen.dart';
+import 'package:stoyco_partners_shared/design/utils/tokens/gen/fonts.gen.dart';
+
+class AnnouncementsContentDescription extends ContentCardDescription {
+  const AnnouncementsContentDescription({super.key, required this.data});
+
+  final AnnouncementsContentModel data;
+
+  @override
+  Widget build(BuildContext context) {
+    return SizedBox(
+      height: StoycoScreenSize.height(context, 115),
+      child: Row(
+        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+        crossAxisAlignment: CrossAxisAlignment.center,
+        children: Gutter.separateChildren(
+          children: <Widget>[
+            Container(
+              constraints: BoxConstraints(
+                maxWidth: StoycoScreenSize.width(context, 146),
+              ),
+              child: Column(
+                mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: Gutter.separateChildren(
+                  children: <Widget>[
+                    Column(
+                      children: <Widget>[
+                        Text(
+                          data.title,
+                          style: TextStyle(
+                            fontFamily: StoycoFontFamilyToken.gilroy,
+                            fontSize: StoycoScreenSize.fontSize(context, 12),
+                            color: ColorFoundation.text.saLight,
+                            fontWeight: FontWeight.w500,
+                          ),
+                          maxLines: 2,
+                          overflow: TextOverflow.ellipsis,
+                        ),
+                        Gutter(StoycoScreenSize.height(context, 5)),
+                        Text(
+                          data.description,
+                          style: TextStyle(
+                            fontFamily: StoycoFontFamilyToken.gilroy,
+                            fontSize: StoycoScreenSize.fontSize(context, 12),
+                            color: ColorFoundation.text.saLight,
+                            fontWeight: FontWeight.w300,
+                          ),
+                          maxLines: 2,
+                          overflow: TextOverflow.ellipsis,
+                        ),
+                      ],
+                    ),
+                    Column(
+                      children: <Widget>[
+                        RichText(
+                          text: TextSpan(
+                            text: 'Participantes: ',
+                            style: TextStyle(
+                              fontFamily: StoycoFontFamilyToken.gilroy,
+                              fontSize: StoycoScreenSize.fontSize(context, 12),
+                              fontWeight: FontWeight.w500,
+                              color: ColorFoundation.text.saHighlights,
+                            ),
+                            children: <InlineSpan>[
+                              TextSpan(
+                                text: ' ${data.participants}',
+                                style: TextStyle(
+                                  fontFamily: StoycoFontFamilyToken.gilroy,
+                                  fontSize: StoycoScreenSize.fontSize(
+                                    context,
+                                    12,
+                                  ),
+                                  fontWeight: FontWeight.w500,
+                                  color: ColorFoundation.text.saLight,
+                                ),
+                              ),
+                            ],
+                          ),
+                        ),
+                        Gutter(StoycoScreenSize.height(context, 5)),
+                        Text(
+                          'Finalizado ${data.closeDate}',
+                          style: TextStyle(
+                            fontFamily: StoycoFontFamilyToken.gilroy,
+                            fontSize: StoycoScreenSize.fontSize(context, 8),
+                            color: ColorFoundation.text.saLight,
+                            fontWeight: FontWeight.w700,
+                          ),
+                        ),
+                      ],
+                    ),
+                  ],
+                ),
+              ),
+            ),
+            Column(
+              mainAxisAlignment: MainAxisAlignment.start,
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: Gutter.separateChildren(
+                children: <Widget>[
+                  ContentStat(
+                    icon: StoycoAssetsToken.lib.assets.icons.like.svg(
+                      package: 'stoyco_partners_shared',
+                      width: StoycoScreenSize.width(context, 14),
+                      height: StoycoScreenSize.height(context, 14),
+                    ),
+                    stat: data.likes.toString(),
+                  ),
+                  ContentStat(
+                    icon: StoycoAssetsToken.lib.assets.icons.share.svg(
+                      package: 'stoyco_partners_shared',
+                      width: StoycoScreenSize.width(context, 14),
+                      height: StoycoScreenSize.height(context, 14),
+                    ),
+                    stat: data.shares.toString(),
+                  ),
+                  ContentStat(
+                    icon: StoycoAssetsToken.lib.assets.icons.message.svg(
+                      package: 'stoyco_partners_shared',
+                      width: StoycoScreenSize.width(context, 14),
+                      height: StoycoScreenSize.height(context, 14),
+                    ),
+                    stat: data.comments.toString(),
+                  ),
+                  ContentStat(
+                    icon: Icon(
+                      Icons.remove_red_eye_outlined,
+                      color: ColorFoundation.text.saLight,
+                      size: StoycoScreenSize.height(context, 14),
+                    ),
+                    stat: data.views.toString(),
+                  ),
+                ],
+                extent: StoycoScreenSize.height(context, 5),
+              ),
+            ),
+          ],
+          extent: StoycoScreenSize.width(context, 18),
+        ),
+      ),
+    );
+  }
+}