+++ lib/design/atomic/organisms/content_card/content_card.dart
@@ -1,7 +1,7 @@
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
-import 'package:stoyco_partners_shared/design/atomic/molecules/content_card_description/content_card_description.dart';
-import 'package:stoyco_partners_shared/design/models/content_card_model.dart';
+import 'package:stoyco_partners_shared/design/atomic/templates/content_card_description.dart';
+import 'package:stoyco_partners_shared/design/models/bases/content_card_model.dart';
 import 'package:stoyco_partners_shared/design/responsive/gutter.dart';
 import 'package:stoyco_partners_shared/design/responsive/screen_size/stoyco_screen_size.dart';
 
@@ -28,8 +28,7 @@ class ContentCard extends StatelessWidget {
                 imageUrl: data.contentImageUrl,
               ),
             ),
-            //Todo: change description type depending on needed content. This is the video/news one. Default style
-            Expanded(child: ContentCardDescription(data: data)),
+            Expanded(child: ContentCardDescription.fromModel(data)),
           ],
           extent: StoycoScreenSize.width(context, 16),
         ),